requirements.txt
fastapi
uvicorn
pydantic
pyyaml
pytest
redis
